interview/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyairtable==1.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.16
